% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getMatch}
\alias{getMatch}
\title{Fetching JSON data for one match}
\usage{
getMatch(id)
}
\arguments{
\item{id}{the unique ID for the match, such as \code{'00cbe304-d7d7-11e6-ad79-062445d3d668'}}
}
\value{
A large data table converted from the JSON response with jsonlite.
    \code{result$data$type} will be the type of object, usually \code{"match"}.
    \code{result$data$id} will be the unique ID of the match, such as
    \code{'00cbe304-d7d7-11e6-ad79-062445d3d668'}.
    \code{result$data$attributes} is a list with various attributes of the match, including its
    time, length, game mode, region, and how it ended.
    \code{result$data$relationships} will be a list of two team rosters, each of which is just an
    ID, of the same format as the match ID.
    \code{result$included} is a table whose rows are players, participants, and rosters,
    in no particular order.  The \code{type} column tells which of the three types the row holds.
    The \code{id} column holds the unique ID for the object in that row.
    The \code{attributes} column holds another data frame; its \code{$actor} column lists the
    names of the heroes for each row that is a player, its \code{$stats} column has copious data
    not documented here about activity throughout the match, and its \code{$name} column has the
    in-game names for each row that is a player.
}
\description{
Fetches the data for one match, given the ID for that match.
This will be a large and complex data structure, described below.
}
\examples{
getMatch( '00cbe304-d7d7-11e6-ad79-062445d3d668' )

}

