% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getMatches}
\alias{getMatches}
\title{Fetching JSON data for many matches}
\usage{
getMatches(pageOffset = 0)
}
\arguments{
\item{pageOffset}{defaults to 0, can be used to page through the huge list of matches}
}
\value{
A large data table converted from the JSON response with jsonlite
}
\description{
Fetches one page of matches from all available matches.
Each match is a very large data structure.
See the documentation for \code{getMatch()} for details.
}
\examples{
getMatches()

}

