% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getPlayer}
\alias{getPlayer}
\title{Fetching JSON data for one player}
\usage{
getPlayer(id)
}
\arguments{
\item{id}{the unique ID for the player, such as \code{'83fc8814-b40a-11e6-8acb-06d90c28bf1a'}}
}
\value{
probably just a 404 error at this point
}
\description{
This does not yet work; the API doesn't support it.  This stub is here for now.
}
\examples{
getPlayer( '83fc8814-b40a-11e6-8acb-06d90c28bf1a' )

}

