% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getPlayers}
\alias{getPlayers}
\title{Fetching JSON data for many players}
\usage{
getPlayers()
}
\value{
probably just a 404 error at this point
}
\description{
This does not yet work; the API doesn't support it.  This stub is here for now.
}
\examples{
getPlayers()

}

