% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getTeam}
\alias{getTeam}
\title{Fetching JSON data for one team}
\usage{
getTeam(id)
}
\arguments{
\item{id}{the unique ID for the player, such as \code{'080a11a6-cfdd-11e6-90c5-0606daf7e67d'}}
}
\value{
probably just a 404 error at this point
}
\description{
This does not yet work; the API doesn't support it.  This stub is here for now.
}
\examples{
getTeam( '080a11a6-cfdd-11e6-90c5-0606daf7e67d' )

}

