% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getTeams}
\alias{getTeams}
\title{Fetching JSON data for many teams}
\usage{
getTeams()
}
\value{
probably just a 404 error at this point
}
\description{
This does not yet work; the API doesn't support it.  This stub is here for now.
}
\examples{
getTeams()

}

